+++ packages/atomic/cypress/e2e/facets/timeframe-facet/timeframe-facet-assertions.ts
@@ -122,3 +122,12 @@ export function assertFacetValueContainsText(index: number, value: string) {
       });
   });
 }
+
+export function assertValuesMatchExpectedOrder(expectedValues: string[]) {
+  it('values should be ordered in the expected order', () => {
+    TimeframeFacetSelectors.valueLabel().as('facetAllValuesLabel');
+    cy.getTextOfAllElements('@facetAllValuesLabel').then((originalValues) => {
+      expect(originalValues).to.eql(expectedValues);
+    });
+  });
+}