+++ packages/atomic/src/components/search/atomic-search-interface/analytics-config.ts
@@ -8,10 +8,25 @@ import {
   augmentWithExternalMiddleware,
   augmentAnalyticsConfigWithDocument,
   augmentAnalyticsConfigWithAtomicVersion,
+  getNextAnalyticsConfig,
 } from '../../common/interface/analytics-config';
 import {createAtomicStore} from './store';
 
 export function getAnalyticsConfig(
+  searchEngineConfig: SearchEngineConfiguration,
+  enabled: boolean,
+  store: ReturnType<typeof createAtomicStore>
+): AnalyticsConfiguration {
+  switch (searchEngineConfig.analytics?.analyticsMode) {
+    case 'next':
+      return getNextAnalyticsConfig(searchEngineConfig, enabled);
+    case 'legacy':
+    default:
+      return getLegacyAnalyticsConfig(searchEngineConfig, enabled, store);
+  }
+}
+
+function getLegacyAnalyticsConfig(
   searchEngineConfig: SearchEngineConfiguration,
   enabled: boolean,
   store: ReturnType<typeof createAtomicStore>