+++ packages/headless/src/controllers/commerce/context/headless-context.ts
@@ -1,3 +1,4 @@
+import {CurrencyCodeISO4217} from '@coveo/relay-event-types';
 import {CommerceEngine} from '../../../app/commerce-engine/commerce-engine';
 import {
   setContext,
@@ -14,10 +15,9 @@ import {
 } from '../../controller/headless-controller';
 
 export interface ContextOptions {
-  trackingId: string;
   language: string;
   country: string;
-  currency: string;
+  currency: CurrencyCodeISO4217;
   user?: User;
   view: View;
 }
@@ -52,12 +52,6 @@ export interface ContextProps {
  */
 export interface Context extends Controller {
   /**
-   * Sets the tracking ID.
-   * @param trackingId - The new tracking ID.
-   */
-  setTrackingId(trackingId: string): void;
-
-  /**
    * Sets the language.
    * @param language - The new language.
    */
@@ -73,7 +67,7 @@ export interface Context extends Controller {
    * Sets the currency.
    * @param currency - The new currency.
    */
-  setCurrency(currency: string): void;
+  setCurrency(currency: CurrencyCodeISO4217): void;
 
   /**
    * Sets the user.
@@ -94,10 +88,9 @@ export interface Context extends Controller {
 }
 
 export interface ContextState {
-  trackingId: string;
   language: string;
   country: string;
-  currency: string;
+  currency: CurrencyCodeISO4217;
   user?: User;
   view: View;
 }
@@ -142,14 +135,6 @@ export function buildContext(
       return getState().commerceContext;
     },
 
-    setTrackingId: (trackingId: string) =>
-      dispatch(
-        setContext({
-          ...getState().commerceContext,
-          trackingId,
-        })
-      ),
-
     setLanguage: (language: string) =>
       dispatch(
         setContext({
@@ -166,7 +151,7 @@ export function buildContext(
         })
       ),
 
-    setCurrency: (currency: string) =>
+    setCurrency: (currency: CurrencyCodeISO4217) =>
       dispatch(
         setContext({
           ...getState().commerceContext,