+++ utils/release/npm-publish-package.mjs
@@ -27,7 +27,9 @@ async function isPublished(name, version, tag = version) {
 }
 
 const isPrerelease = process.env.IS_PRERELEASE === 'true';
-
+const shouldProvideProvenance =
+  !isPrerelease &&
+  process.env.npm_config_registry !== 'https://npm.pkg.github.com';
 /** @type {import('@npmcli/package-json').PackageJson} */
 const {name, version} = JSON.parse(
   readFileSync('package.json', {encoding: 'utf-8'})
@@ -37,7 +39,10 @@ if (!name || !version) {
 }
 if (!(await isPublished(name, version))) {
   const tagToPublish = isPrerelease ? 'alpha' : 'beta';
-  await npmPublish('.', {tag: tagToPublish});
+  await npmPublish('.', {
+    tag: tagToPublish,
+    provenance: shouldProvideProvenance,
+  });
   await retry(
     async () => {
       if (!(await isPublished(name, version, tagToPublish))) {